Based on the provided code changes, I would suggest the following Git commit message:
`refactor(constants/ErrorMessage): Update error messages`

This commit message follows the Conventional Commits 1.0.0 specification, using the `refactor` type to indicate a code change that neither fixes a bug nor adds a feature.

Note: The original file was deleted from the main branch and replaced with a new version of `ErrorMessage.java`. However, since this is not a traditional refactor (as it involves deleting and replacing files), I've chosen a different commit message that still accurately describes the changes.